Moved help command into separate file to be shared by the CLI and Command constructors.
utils transformValue() will try to guess the value type if a type is not specified.

allowUnknownOptions has been renamed to treatUnknownOptionsAsArguments and negated. Unknown options are always stored in argv, but only when treatUnknownOptionsAsArguments is true does it add it to _.

Reworking the help to generate an object with all the help data.