Finished generating help as an object.
Added support to options and arguments for regular expressions as a 'type'.

Added support to options for 'count' types.

Fixed bug with the default command not firing.

Removed style() function from Context.

Added support for a generic errorMsg to options.

Added a default help renderer function with the ability to override.

Removed debug code.

Added hooks for parsing and generating help.

Fixed support for custom ordering of options.

Tons of misc code cleanup.